Update mypy ci and pyprojects and minor corrections

+++ geos-posp/src/geos_posp/filters/SurfaceGeomechanics.py
@@ -428,7 +428,7 @@ class SurfaceGeomechanics(VTKPythonAlgorithmBase):
         if fromLocalToYXZ:
             return P
         # inverse the change of basis matrix
-        return np.linalg.inv(P)
+        return np.linalg.inv(P).astype(np.float64)
 
     def getNormalTangentsVectors(self: Self) -> npt.NDArray[np.float64]:
         """Compute the change of basis matrix from Local to XYZ bases.
@@ -447,7 +447,7 @@ class SurfaceGeomechanics(VTKPythonAlgorithmBase):
         assert tangents1 is not None, "Tangents attribute was not found."
 
         # compute second tangential component
-        tangents2: npt.NDArray[np.float64] = np.cross(normals, tangents1, axis=1)
+        tangents2: npt.NDArray[np.float64] = np.cross(normals, tangents1, axis=1).astype(np.float64)
         assert tangents2 is not None, "Local basis third axis was not computed."
 
         # put vectors as columns
